refactor: refactor scritps
