--- src/gen_code/common/constants.py
@@ -1,3 +1,0 @@
-from pathlib import Path
-
-ROOT_DIR = Path(__file__).resolve().parent.parent.parent.parent
